backend/requirements.txt
Flask
Werkzeug
Jinja2
Flask-SQLAlchemy
redis
celery
python-dotenv
itsdangerous
PyJWT